+++ src/main/java/com/verdantartifice/primalmagick/datagen/loot_tables/BlockLootTables.java
@@ -75,6 +75,7 @@ public class BlockLootTables extends AbstractBlockLootTableProvider {
         this.registerManaBearingDeviceTable(BlocksPM.ESSENCE_TRANSMUTER.get());
         this.registerManaBearingDeviceTable(BlocksPM.DISSOLUTION_CHAMBER.get());
         this.registerBasicTable(BlocksPM.ZEPHYR_ENGINE.get());
+        this.registerBasicTable(BlocksPM.VOID_TURBINE.get());
 
         // Register misc loot tables
         this.registerBasicTable(BlocksPM.SALT_TRAIL.get());