+++ src/main/java/com/verdantartifice/primalmagick/common/init/InitStats.java
@@ -48,9 +48,9 @@ public class InitStats {
         StatTriggers.register(StatsPM.RITUALS_COMPLETED, 25, SimpleResearchKey.parse("t_rituals_completed_master"));
         StatTriggers.register(StatsPM.RITUALS_COMPLETED, 100, SimpleResearchKey.parse("t_rituals_completed_supreme"));
         StatTriggers.register(StatsPM.RITUAL_MISHAPS, 1, SimpleResearchKey.parse("t_ritual_mishaps_basic"));
-        StatTriggers.register(StatsPM.CRAFTED_MAGITECH, 10, SimpleResearchKey.parse("b_crafted_magitech_expert"));
-        StatTriggers.register(StatsPM.CRAFTED_MAGITECH, 50, SimpleResearchKey.parse("b_crafted_magitech_master"));
-        StatTriggers.register(StatsPM.CRAFTED_MAGITECH, 250, SimpleResearchKey.parse("b_crafted_magitech_supreme"));
+        StatTriggers.register(StatsPM.CRAFTED_MAGITECH, 5, SimpleResearchKey.parse("b_crafted_magitech_expert"));
+        StatTriggers.register(StatsPM.CRAFTED_MAGITECH, 25, SimpleResearchKey.parse("b_crafted_magitech_master"));
+        StatTriggers.register(StatsPM.CRAFTED_MAGITECH, 100, SimpleResearchKey.parse("b_crafted_magitech_supreme"));
         StatTriggers.register(StatsPM.ITEMS_RUNESCRIBED, 2, SimpleResearchKey.parse("t_items_runescribed_expert"));
         StatTriggers.register(StatsPM.ITEMS_RUNESCRIBED, 10, SimpleResearchKey.parse("t_items_runescribed_master"));
         StatTriggers.register(StatsPM.ITEMS_RUNESCRIBED, 50, SimpleResearchKey.parse("t_items_runescribed_supreme"));